+++ lib/Processor.php
@@ -5,7 +5,7 @@ namespace SearchEngine;
 /**
  * SearchEngine Processor
  *
- * @version 0.1.1
+ * @version 0.1.2
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
  */
@@ -20,7 +20,7 @@ class Processor extends Base {
      * @param array $index Index as an array.
      * @return string Processed index string.
      */
-    protected function processIndex(array $index): string {
+    public function processIndex(array $index): string {
         $processed_index = '';
         if (!empty($index)) {
             $meta_index = [];